move (lazyly) mixins to devices
