+++ custom_components/meross_lan/number.py
@@ -6,7 +6,6 @@ from homeassistant.components import number
 from homeassistant.const import PERCENTAGE, TEMP_CELSIUS
 
 from . import meross_entity as me
-from .helpers import SmartPollingStrategy
 from .merossclient import const as mc, get_namespacekey  # mEROSS cONST
 
 if typing.TYPE_CHECKING:
@@ -161,93 +160,3 @@ class MLHubAdjustNumber(MLConfigNumber):
     @property
     def ml_multiplier(self):
         return 100
-
-
-class MLScreenBrightnessNumber(MLConfigNumber):
-    manager: ScreenBrightnessMixin
-
-    _attr_icon = "mdi:brightness-percent"
-
-    def __init__(self, manager: ScreenBrightnessMixin, channel: object, key: str):
-        self.key_value = key
-        self._attr_name = f"Screen brightness ({key})"
-        super().__init__(manager, channel, f"screenbrightness_{key}")
-
-    @property
-    def native_max_value(self):
-        return 100
-
-    @property
-    def native_min_value(self):
-        return 0
-
-    @property
-    def native_step(self):
-        return 12.5
-
-    @property
-    def native_unit_of_measurement(self):
-        return PERCENTAGE
-
-    async def async_set_native_value(self, value: float):
-        brightness = {
-            mc.KEY_CHANNEL: self.channel,
-            mc.KEY_OPERATION: self.manager._number_brightness_operation.native_value,
-            mc.KEY_STANDBY: self.manager._number_brightness_standby.native_value,
-        }
-        brightness[self.key_value] = value
-
-        def _ack_callback(acknowledge: bool, header: dict, payload: dict):
-            if acknowledge:
-                self.update_native_value(value)
-
-        await self.manager.async_request(
-            mc.NS_APPLIANCE_CONTROL_SCREEN_BRIGHTNESS,
-            mc.METHOD_SET,
-            {mc.KEY_BRIGHTNESS: [brightness]},
-            _ack_callback,
-        )
-
-
-class ScreenBrightnessMixin(
-    MerossDevice if typing.TYPE_CHECKING else object
-):  # pylint: disable=used-before-assignment
-    _number_brightness_operation: MLScreenBrightnessNumber
-    _number_brightness_standby: MLScreenBrightnessNumber
-
-    def __init__(self, descriptor, entry):
-        super().__init__(descriptor, entry)
-
-        with self.exception_warning("ScreenBrightnessMixin init"):
-            # the 'ScreenBrightnessMixin' actually doesnt have a clue of how many  entities
-            # are controllable since the digest payload doesnt carry anything (like MerossShutter)
-            # So we're not implementing _init_xxx and _parse_xxx methods here and
-            # we'll just add a couple of number entities to control 'active' and 'standby' brightness
-            # on channel 0 which will likely be the only one available
-            self._number_brightness_operation = MLScreenBrightnessNumber(
-                self, 0, mc.KEY_OPERATION
-            )
-            self._number_brightness_standby = MLScreenBrightnessNumber(
-                self, 0, mc.KEY_STANDBY
-            )
-            self.polling_dictionary[
-                mc.NS_APPLIANCE_CONTROL_SCREEN_BRIGHTNESS
-            ] = SmartPollingStrategy(mc.NS_APPLIANCE_CONTROL_SCREEN_BRIGHTNESS)
-
-    # interface: MerossDevice
-    async def async_shutdown(self):
-        await super().async_shutdown()
-        self._number_brightness_operation = None  # type: ignore
-        self._number_brightness_standby = None  # type: ignore
-
-    # interface: self
-    def _handle_Appliance_Control_Screen_Brightness(self, header: dict, payload: dict):
-        for p_channel in payload[mc.KEY_BRIGHTNESS]:
-            if p_channel.get(mc.KEY_CHANNEL) == 0:
-                self._number_brightness_operation.update_native_value(
-                    p_channel[mc.KEY_OPERATION]
-                )
-                self._number_brightness_standby.update_native_value(
-                    p_channel[mc.KEY_STANDBY]
-                )
-                break
